refactor(main): extract useBeltColorLabel to reusable hook
